Feature/seed update

See merge request entelect-challenge/ec-dev-infrastructure/game-engine-interface!9

+++ src/main/java/za/co/entelect/challenge/game/contracts/bootstrapper/GameEngineBootstrapper.java
@@ -23,7 +23,7 @@ import za.co.entelect.challenge.game.contracts.renderer.RendererType;
  */
 public interface GameEngineBootstrapper {
 
-    void setSeed(long seed);
+    void setSeed(int seed);
 
     void setConfigPath(String path);
